+++ Konsole/MockConsole.cs
@@ -14,6 +14,8 @@ namespace Konsole
     public class MockConsole : Window
     {
         public MockConsole(int x, int y, params K[] options) : base(x, y, null, options) { }
+        public MockConsole(int width, int height, ConsoleColor foreground, ConsoleColor background, params K[] options) 
+            : base(0,0, width, height, foreground, background, false, null, options) { }
         public MockConsole() : base(120, 60, false) { }
         public MockConsole(int width, int height) : base(width, height, false) { }